Revert "feat(DTFS2-6892): update mock builder to require complete defaults, a…"
This reverts commit 6f04dad88f1067a8287768c740019156ce294f91.
