+++ libs/common/src/test-helpers/mock-builders/index.ts
@@ -1,1 +1,1 @@
-export * from './mock-builder';
+export * from './mock-builder.mock.builder';
