+++ trade-finance-manager-api/src/v1/__mocks__/builders/user.service.mock.builder.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { BaseMockBuilder, EntraIdUser, UpsertTfmUserRequest } from '@ukef/dtfs2-common';
+import { BaseMockBuilder } from '@ukef/dtfs2-common';
 import { aTfmUser } from '@ukef/dtfs2-common/mock-data-backend';
 import {
   UpsertTfmUserFromEntraIdUserParams,
@@ -10,16 +10,8 @@ import {
 
 export class UserServiceMockBuilder extends BaseMockBuilder<UserService> {
   constructor() {
-    const userService = new UserService();
     super({
       defaultInstance: {
-        /**
-         *  We pass through the actual service implementation for the below method here,
-         * as it is an existing pattern that we never mock synchronous methods.
-         */
-        transformEntraIdUserToUpsertTfmUserRequest(entraIdUser: EntraIdUser): UpsertTfmUserRequest {
-          return userService.transformEntraIdUserToUpsertTfmUserRequest(entraIdUser);
-        },
         upsertTfmUserFromEntraIdUser({ entraIdUser, auditDetails }: UpsertTfmUserFromEntraIdUserParams): Promise<UpsertTfmUserFromEntraIdUserResponse> {
           return Promise.resolve(aTfmUser());
         },
