+++ trade-finance-manager-api/src/v1/controllers/sso.controller.get-auth-code-url.test.ts
@@ -31,8 +31,8 @@ describe('SsoController', () => {
 
     beforeEach(() => {
       jest.resetAllMocks();
-      entraIdService = new EntraIdServiceMockBuilder().with({ getAuthCodeUrl: getAuthCodeUrlMock }).build();
-      userService = new UserServiceMockBuilder().build();
+      entraIdService = new EntraIdServiceMockBuilder().withDefaults().with({ getAuthCodeUrl: getAuthCodeUrlMock }).build();
+      userService = new UserServiceMockBuilder().withDefaults().build();
       ssoController = new SsoController({ entraIdService, userService });
 
       ({ req, res } = getHttpMocks());
