+++ trade-finance-manager-api/src/v1/controllers/sso.controller.handle-sso-redirect-form.test.ts
@@ -38,8 +38,9 @@ describe('SsoController', () => {
 
     beforeEach(() => {
       jest.resetAllMocks();
-      entraIdService = new EntraIdServiceMockBuilder().with({ handleRedirect: handleRedirectMock }).build();
+      entraIdService = new EntraIdServiceMockBuilder().withDefaults().with({ handleRedirect: handleRedirectMock }).build();
       userService = new UserServiceMockBuilder()
+        .withDefaults()
         .with({ upsertTfmUserFromEntraIdUser: upsertTfmUserFromEntraIdUserMock, saveUserLoginInformation: saveUserLoginInformationMock })
         .build();
       ssoController = new SsoController({ entraIdService, userService });
